+++ includes/class-stancer.php
@@ -12,6 +12,9 @@
  * @subpackage stancer/includes
  */
 
+use Automattic\WooCommerce\Blocks\Integrations\IntegrationRegistry;
+use Automattic\WooCommerce\Blocks\Payments\PaymentMethodRegistry;
+
 /**
  * Stancer plugin.
  *
@@ -82,7 +85,6 @@ class WC_Stancer {
 		require_once ABSPATH . 'wp-admin/includes/upgrade.php';
 
 		// phpcs:disable WordPress.DB.PreparedSQL.NotPrepared
-
 		if ( version_compare( '1.0.0', $version, '>' ) ) {
 			$sql = 'CREATE TABLE ' . $wpdb->prefix . 'wc_stancer_card (
 				stancer_card_id int(10) UNSIGNED NOT NULL AUTO_INCREMENT COMMENT "Unique ID in this table",
@@ -171,6 +173,14 @@ class WC_Stancer {
 		add_action( 'plugins_loaded', [ $this, 'load_plugin' ] );
 		add_action( 'wc_ajax_create_order', [ $this, 'create_order' ] );
 		add_action( 'wp_enqueue_scripts', [ $this, 'load_public_hooks' ] );
+		add_action( 'woocommerce_blocks_loaded', [ $this, 'gateway_block_support' ] );
+		add_action(
+			'rest_api_init',
+			function () {
+				$payment_change_controller = new WCS_Stancer_Change_Payment_Method();
+				$payment_change_controller->register_routes();
+			}
+		);
 		add_action( 'admin_notices', [ $this, 'display_depreciation' ] );
 	}
 
@@ -245,6 +255,33 @@ class WC_Stancer {
 	public function load_plugin() {
 		$this->upgrade_plugin();
 		$this->load_gateway();
+	}
+	/**
+	 * Register the Gateway Block support to the approriate action hooks
+	 *
+	 * @since 1.3.0
+	 *
+	 * @return void
+	 */
+	public function gateway_block_support() {
+		if ( class_exists( 'Automattic\WooCommerce\Blocks\Payments\Integrations\AbstractPaymentMethodType' ) ) {
+			add_action(
+				'woocommerce_blocks_payment_method_type_registration',
+				function ( PaymentMethodRegistry $payment_method_registry ) {
+					$payment_method_registry->register( new WC_Stancer_Gateway_Block_Support() );
+				}
+			);
+		}
+	}
+	/**
+	 * Get the plugin absolute path
+	 *
+	 * @since 1.3.0
+	 *
+	 * @return string
+	 */
+	public function plugin_abspath() {
+		return plugin_dir_path( dirname( ( __FILE__ ) ) );
 	}
 
 	/**