+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2024-09-26
+
+### Added
+- Support of woocommerce checkout blocks. (CMS-68)
+
+### Changed
+- Change the cb logo to the new logo (CMS-173)
+
+### Fixed
+- Fix plugin sometimes being active when is shouldn't (CMS-171)
+- Fix payment description not showing correctly on redirect mode (CMS-137)
+
+
 ## [1.2.6] - 2024-08-08
 
 ### Added